tests/wpt/tests/paint-timing/first-contentful-canvas-webgl2.html

<!DOCTYPE html>

<head>
  <title>Performance Paint Timing Test: FCP due to canvas</title>
</head>

<body>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="resources/utils.js"></script>
  <canvas id="canvas" width="200" height="200"></canvas>

  <script>
    setup({ "hide_test_state": true });
    async_test(function (t) {
      assert_implements(window.PerformancePaintTiming, "Paint Timing isn't supported.");
      const canvas = document.getElementById("canvas");
      const context = canvas.getContext("webgl2");
      if (!context) {
        assert_implements_optional(context, "WebGL 2 Canvas isn't supported.")
      }
      context.clearColor(0.3, 0.3, 0.3, 1);
      context.clear(context.COLOR_BUFFER_BIT);
      assertPaintTimingEntries(t, 20);
    }, "First contentful paint fires due to webgl2 canvas render.");
  </script>
</body>

</html>
